feat(movie-service): refactor CinemaRoom and Seat entities to match DB v2

See merge request hcm26_cpl_java_05/hcm26_cpl_java_05_group1!113

+++ server/movie-service/src/main/java/movieservice/repository/CinemaRoomRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import movieservice.entity.CinemaRoom;
 
-public interface CinemaRoomRepository extends JpaRepository<CinemaRoom, Integer> {
+public interface CinemaRoomRepository extends JpaRepository<CinemaRoom, Long> {
     CinemaRoom findByCinemaRoomId(Long cinemaId);
     boolean existsByCinemaRoomName(String cinemaRoomName);
 }
